requirements.txt
numpy==1.21.0
torch==1.10.0
torchvision==0.11.1
transformers==4.11.3
scikit-learn==0.24.2
Pillow==8.4.0
matplotlib==3.4.3
tqdm==4.62.3
yaml==5.4.1
opencv-python==4.5.3.20210927